% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterPDBbfactor.R
\name{getPDBinfo}
\alias{getPDBinfo}
\title{getPDBinfo.R}
\usage{
getPDBinfo(pdbid, uniprotAcc)
}
\arguments{
\item{pdbid}{a PDB identifier (e.g. 2GS6) for only one PDB or the word "all" for all the available PDBs that correspond to the UniProtAC}

\item{uniprotAcc}{the uniProtKB accession code of your protein (e.g. P00533)}
}
\value{
A dataframe with PDB information
}
\description{
gets the information of the PDB of interest, and identifies those regions that have missing coordinates
}
