% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterPDBbfactor.R
\name{alterPDBbfactor}
\alias{alterPDBbfactor}
\title{alterPDBbfactor.R}
\usage{
alterPDBbfactor(pdbid, uniprotAcc, seq_scores, outfolder)
}
\arguments{
\item{pdbid}{a PDB identifier (e.g. 2GS6)}

\item{uniprotAcc}{the uniProtKB accession code of your protein (e.g. P00533)}

\item{seq_scores}{A dataframe with two columns: residue position (aacPos) number and the by-residue scores (score) you want to add in the B-factor field of the PDB (e.g. conservation)
aacPos | score
1      |  0.1
2      |  1
3      |  0.5}

\item{outfolder}{The output folder}

\item{chain_id}{a chain of yoyr PDB of interest (e.g. A)}
}
\value{
The PyMol script
}
\description{
Download and renumber a PDB structure, then alters the B-factor column replacing by scores/numbers of your interest
}
